--- components/draggables/index.ts
@@ -1,1 +1,0 @@
-export { default as DraggableItem } from './DraggableItem';
